reorganize code in solverInterface directory
